requirements.txt
cyhunspell==2.0.2
lxml==4.6.3
requests==2.26.0
difflib
